fix form reviews
