+++ nhost/metadata/databases/default/tables/public_log.yaml
@@ -78,6 +78,7 @@ update_permissions:
     permission:
       columns:
         - canceled
+        - threadId
       filter:
         org:
           members:
